partial pull-request from jfehr, ommiting changes to openssl lib names: Changes needed to build Cmam in VS2015.

+++ libclamav/pdf.c
@@ -908,18 +908,18 @@ int pdf_extract_obj(struct pdf_struct *pdf, struct pdf_obj *obj, uint32_t flags)
                     if (length < 0)
                         length = 0;
 
-                    cli_dbgmsg("cli_pdf: calculated length %jd\n", (intmax_t)length);
+                    cli_dbgmsg("cli_pdf: calculated length %jd\n", length);
                 } else {
                     if (size > (size_t)length+2) {
                         cli_dbgmsg("cli_pdf: calculated length %jd < %jd\n",
-                                   (intmax_t)length, (intmax_t)size);
+                                   length, size);
                         length = size;
                     }
                 }
 
                 if (orig_length && size > (size_t)orig_length + 20) {
                     cli_dbgmsg("cli_pdf: orig length: %jd, length: %jd, size: %jd\n",
-                               (intmax_t)orig_length, (intmax_t)length, (intmax_t)size);
+                               orig_length, length, size);
                     pdfobj_flag(pdf, obj, BAD_STREAMLEN);
                 }
 
@@ -1055,7 +1055,7 @@ int pdf_extract_obj(struct pdf_struct *pdf, struct pdf_obj *obj, uint32_t flags)
         }
     } while (0);
 
-    cli_dbgmsg("cli_pdf: extracted %jd bytes %u %u obj\n", (intmax_t)sum, obj->id>>8, obj->id&0xff);
+    cli_dbgmsg("cli_pdf: extracted %jd bytes %u %u obj\n", sum, obj->id>>8, obj->id&0xff);
     cli_dbgmsg("         ... to %s\n", fullname);
 
     if (flags & PDF_EXTRACT_OBJ_SCAN && sum) {
@@ -2320,7 +2320,7 @@ int cli_pdf(const char *dir, cli_ctx *ctx, off_t offset)
 
     if (pdfver != start || offset) {
         pdf.flags |= 1 << BAD_PDF_HEADERPOS;
-        cli_dbgmsg("cli_pdf: PDF header is not at position 0: %jd\n",(intmax_t)(pdfver-start+offset));
+        cli_dbgmsg("cli_pdf: PDF header is not at position 0: %jd\n",(pdfver-start+offset));
 #if HAVE_JSON
         if (pdfobj)
             cli_jsonbool(pdfobj, "BadVersionLocation", 1);
@@ -2429,7 +2429,7 @@ int cli_pdf(const char *dir, cli_ctx *ctx, off_t offset)
     while ((rc = pdf_findobj(&pdf)) > 0) {
         struct pdf_obj *obj = &pdf.objs[pdf.nobjs-1];
 
-        cli_dbgmsg("cli_pdf: found %d %d obj @%jd\n", obj->id >> 8, obj->id&0xff, (intmax_t)(obj->start + offset));
+        cli_dbgmsg("cli_pdf: found %d %d obj @%jd\n", obj->id >> 8, obj->id&0xff, (obj->start + offset));
     }
 
     if (pdf.nobjs)
